+++ Source/Classes/View/CardBalance/CardBalance.swift
@@ -81,9 +81,11 @@ public class CardBalance: UIView {
             return
         }
         
-        balanceTitle.font = UIFont.systemFont(ofSize: titleFontSize, weight: .regular)
-        balanceTitle.text = model.title.message
-        
+        if let title = model.title {
+            balanceTitle.font = UIFont.systemFont(ofSize: titleFontSize, weight: .regular)
+            balanceTitle.text = title.message
+            setupLabelColors(balanceTitle, field: title)
+        }
         balanceLabel.font = UIFont.systemFont(ofSize: balanceFontSize, weight: .semibold)
         
         eyeImage = UIImageView()
@@ -96,8 +98,6 @@ public class CardBalance: UIView {
         balanceLabel.sizeToFit()
         balanceTitle.sizeToFit()
         eyeImage.sizeToFit()
-        
-        setupLabelColors(balanceTitle, field: model.title)
 
         self.addSubview(balanceTitle)
         self.addSubview(balanceLabel)